Refactor commands and subcommands into `cmds` and `cmds::sub` modules.
