+++ crate/rt/src/cmds.rs
@@ -1,19 +1,12 @@
 pub use self::{
-    diff_cmd::DiffCmd, ensure_cmd::EnsureCmd,
-    states_current_discover_cmd::StatesCurrentDiscoverCmd,
-    states_current_display_cmd::StatesCurrentDisplayCmd,
-    states_current_read_cmd::StatesCurrentReadCmd,
-    states_desired_discover_cmd::StatesDesiredDiscoverCmd,
-    states_desired_display_cmd::StatesDesiredDisplayCmd,
-    states_desired_read_cmd::StatesDesiredReadCmd, states_discover_cmd::StatesDiscoverCmd,
+    diff_cmd::DiffCmd, ensure_cmd::EnsureCmd, states_current_display_cmd::StatesCurrentDisplayCmd,
+    states_desired_display_cmd::StatesDesiredDisplayCmd, states_discover_cmd::StatesDiscoverCmd,
 };
+
+pub mod sub;
 
 mod diff_cmd;
 mod ensure_cmd;
-mod states_current_discover_cmd;
 mod states_current_display_cmd;
-mod states_current_read_cmd;
-mod states_desired_discover_cmd;
 mod states_desired_display_cmd;
-mod states_desired_read_cmd;
 mod states_discover_cmd;
