+++ crate/rt/src/cmds/ensure_cmd.rs
@@ -15,7 +15,7 @@ use peace_rt_model::{
     CmdContext, Error, FnRef, ItemSpecBoxed, ItemSpecGraph, OutputWrite, StatesTypeRegs,
 };
 
-use crate::cmds::{DiffCmd, StatesCurrentDiscoverCmd};
+use crate::cmds::{sub::StatesCurrentDiscoverCmd, DiffCmd};
 
 #[derive(Debug)]
 pub struct EnsureCmd<E, O>(PhantomData<(E, O)>);
